chore: update supported node versions
As a library this should support multiple Nextcloud versions which might
still use Node 20 (Nextcloud < 31) or Node 22 (Nextcloud 31) or even
Node 24 which would be Nextcloud 32 or 33.
This package is not bound to Ndoe as this is a browser library, so we
can even support all LTS versions.

Instead we only should ensure we have a consistent development engine.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>